Integrate Python override regressions into existing input tests
